server/requirements.txt
Django==4.2.3
django-cors-headers==4.2.0
django-ninja==0.22.2
channels==4.0.0
channels-redis==4.1.0
asgiref==3.7.2
async-timeout==4.0.3
daphne==4.0.0
psycopg2-binary==2.9.1